code/backend/requirements.txt
azure-identity~=1.17.1
azure-monitor-opentelemetry~=1.6.2
botbuilder-core~=4.16.1
botbuilder-azure~=4.16.1
botbuilder-integration-aiohttp~=4.16.1
botbuilder-dialogs~=4.16.1
openai~=1.44.1
opentelemetry-instrumentation-aiohttp-client~=0.48b0
opentelemetry-instrumentation-aiohttp-server~=0.48b0
pydantic~=2.9.0
pydantic-settings~=2.4.0
